DON-704: Move login out of donation start primary component
