+++ src/app/donation-start/donation-start-primary/donation-start-primary.component.html
@@ -3,16 +3,6 @@
       donation-start-container.component, along with support code in the .component.ts file, tests etc. It should be
       possible to move piece by piece from the top start or end of the div.
     -->
-<app-donation-start-login
-[logout]="logout"
-[personId]="personId"
-[email]="paymentGroup.value.emailAddress"
-[personIsLoginReady]="personIsLoginReady"
-[loadAuthedPersonInfo]="loadAuthedPersonInfo"
-[creditPenceToUse]="creditPenceToUse"
-[campaign]="campaign"
-[canLogin]="canLogin">
-</app-donation-start-login>
 
 <p *ngIf="!campaignOpenOnLoad">Donations open {{ campaign.startDate | date : 'h:mm a, d LLLL yyyy' }} to {{ campaign.endDate | date : 'h:mm a, d LLLL yyyy' }}</p>
 <p *ngIf="reservationExpiryTime()" class="c-timer">
